+++ tests/vector.rs
@@ -1,9 +1,9 @@
 use sized_matrix::Vector;
-use init_trait::Init;
+use higher_order_functions::Init;
 
 #[test]
 fn init_vector() {
-	let vector = Vector::<usize, 4>::init(|x: usize| x);
+	let vector = Vector::<_, 4>::init(|x: usize| x);
 	
 	for i in 0..4 {
 		assert_eq!(vector[i], i);
@@ -14,5 +14,5 @@ fn init_vector() {
 fn matrix_vector() {
 	let vector = Vector::vector([0, 1, 2, 3]);
 	
-	assert_eq!(vector, Vector::<usize, 4>::init(|x: usize| x));
+	assert_eq!(vector, Vector::<_, 4>::init(|x: usize| x));
 }