+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.2.0](https://crates.io/crates/sized_matrix/0.2.0) - 2020-09-04
+
+### Changed:
+* Moved from [init_trait](https://crates.io/crates/init_trait) to [higher_order_functions](https://crates.io/crates/higher_order_functions)
+
+### Added:
+* Implemented [Map](https://docs.rs/higher_order_functions/0.1.0/higher_order_functions/trait.Map.html) for matrices
+
 ## [v0.1.0](https://crates.io/crates/sized_matrix/0.1.0) - 2020-08-23
 
 ### Added: